feat: simple grid spacingX and spacingY
