+++ src/components/composites/SimpleGrid/index.tsx
@@ -1,0 +1,2 @@
+export { SimpleGrid } from './SimpleGrid';
+export { ISimpleGridProps } from './props';
