+++ src/components/composites/SimpleGrid/props.ts
@@ -1,7 +1,9 @@
-import type { IBoxProps } from '../Box';
+import type { IBoxProps } from '@native-base/v3';
 
 export type ISimpleGridProps = IBoxProps & {
   spacing?: number;
   columns?: number;
   minChildWidth?: number;
+  spacingX?: number;
+  spacingY?: number;
 };
